Some Changes
I have changed the order of the containers as well as removed some unwanted stuffs.You use the images which are not used now in the photo gallery. Also remove all the contents of the registration container, we only need to put up a link to the registration so a button would suffice. Also work on the preloader.